test(harness): move live Responses API tests to responses_api_live suite
Per audit 5/8c (the scope-doc keeper, migrating as harness cells):
- test_openai_responses_api.py: TestOpenAIResponsesAPITest + the 15 live
  tests (streaming/non-streaming logging, headers, stream validation, router
  + router streaming, bad-request pair, MCP tools, field types, websearch,
  token-limit error, streaming logging, compact) moved with the
  TestCustomLogger/validate_* helpers they alone used; the o1-pro mock
  goldens, store-field transformation and router-no-metadata keepers stay;
  the module-level MockResponse/extra_body fixture pair died with the
  already-deleted extra-body drops
- test_anthropic_responses_api.py: TestAnthropicResponsesAPITest +
  test_multiturn_tool_calls moved; file deleted (nothing remained after the
  drop commit)
- test_azure_responses_api.py: TestAzureResponsesAPITest + preview
  api-version test moved; status-stripping and header-prefix goldens stay
- test_google_ai_studio_responses_api.py: live tools + thought-signature
  tests + base subclass moved; the bridge-transform mock golden stays
