+++ tests/llm_responses_api_testing/test_azure_responses_api.py
@@ -1,54 +1,12 @@
 import os
 import sys
 import pytest
-import asyncio
-from typing import Optional
-from unittest.mock import patch, AsyncMock
+from unittest.mock import patch
 
 sys.path.insert(0, os.path.abspath("../.."))
 import litellm
-from litellm.integrations.custom_logger import CustomLogger
 import json
-from litellm.types.utils import StandardLoggingPayload
-from litellm.types.llms.openai import (
-    ResponseCompletedEvent,
-    ResponsesAPIResponse,
-    ResponseAPIUsage,
-    IncompleteDetails,
-)
 from litellm.llms.custom_httpx.http_handler import AsyncHTTPHandler
-from base_responses_api import BaseResponsesAPITest
-
-
-class TestAzureResponsesAPITest(BaseResponsesAPITest):
-    def get_base_completion_call_args(self):
-        return {
-            "model": "azure/gpt-4.1-mini",
-            "truncation": "auto",
-            "api_base": os.getenv("AZURE_AI_API_BASE"),
-            "api_key": os.getenv("AZURE_AI_API_KEY"),
-            "api_version": "2025-03-01-preview",
-        }
-
-    def get_advanced_model_for_shell_tool(self) -> Optional[str]:
-        """If specified, overrides the model used by test_responses_api_shell_tool_streaming_sees_shell_output (e.g. openai/gpt-5.2 for shell support)."""
-        return "azure/gpt-5-mini"
-
-
-@pytest.mark.asyncio
-async def test_azure_responses_api_preview_api_version():
-    """
-    Ensure new azure preview api version is working
-    """
-    litellm._turn_on_debug()
-    response = await litellm.aresponses(
-        model="azure/gpt-5-mini",
-        truncation="auto",
-        api_version="preview",
-        api_base=os.getenv("AZURE_AI_API_BASE"),
-        api_key=os.getenv("AZURE_AI_API_KEY"),
-        input="Hello, can you tell me a short joke?",
-    )
 
 
 @pytest.mark.asyncio
@@ -57,7 +15,6 @@ async def test_azure_responses_api_status_error():
     Test that 'status' field is not sent in the final request body to Azure API.
     The status field should be filtered out from input messages before making the API call.
     """
-    from unittest.mock import AsyncMock, MagicMock
     import json
 
     request_data = {
@@ -198,7 +155,6 @@ async def test_azure_responses_api_headers_with_llm_provider_prefix():
     in response._hidden_params["headers"] instead of additional_headers, making them
     accessible via completion.headers in the same way as the completion API.
     """
-    import json
     import httpx
 
     mock_response_data = {
